# cargo-ship-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/equipment": {
      "post": {
        "description": "提供设备的基本信息来创建一条新的设备记录。",
        "operationId": "EquipmentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEquipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "设备创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "400": {
            "description": "请求参数验证失败"
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "409": {
            "description": "设备编号已存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建新设备",
        "tags": [
          "设备管理"
        ]
      },
      "get": {
        "description": "分页查询设备列表，支持按类型、状态和关键词进行筛选。",
        "operationId": "EquipmentController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "deviceType",
            "required": false,
            "in": "query",
            "description": "设备类型筛选",
            "schema": {
              "example": "主机",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "设备状态筛选",
            "schema": {
              "example": "normal",
              "type": "string",
              "enum": [
                "normal",
                "warning",
                "fault",
                "offline"
              ]
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词搜索（设备编号或名称）",
            "schema": {
              "example": "ENG",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询设备列表",
        "tags": [
          "设备管理"
        ]
      }
    },
    "/api/equipment/statistics": {
      "get": {
        "description": "获取所有设备按状态（正常、告警、故障、离线）分类的数量统计。",
        "operationId": "EquipmentController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取设备统计信息",
        "tags": [
          "设备管理"
        ]
      }
    },
    "/api/equipment/{id}": {
      "get": {
        "description": "根据设备的唯一ID查询其详细信息。",
        "operationId": "EquipmentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "设备ID (UUID格式)",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "设备不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询设备详情",
        "tags": [
          "设备管理"
        ]
      },
      "patch": {
        "description": "更新指定设备的部分或全部信息。",
        "operationId": "EquipmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "设备ID (UUID格式)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEquipmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "设备不存在"
          },
          "409": {
            "description": "设备编号与现有其他设备冲突"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新设备信息",
        "tags": [
          "设备管理"
        ]
      },
      "delete": {
        "description": "软删除指定设备，记录将保留在数据库中但无法通过常规查询获取。",
        "operationId": "EquipmentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "设备ID (UUID格式)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "设备不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除设备",
        "tags": [
          "设备管理"
        ]
      }
    },
    "/api/equipment/{id}/status": {
      "patch": {
        "description": "快速更新指定设备的运行状态。",
        "operationId": "EquipmentController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "设备ID (UUID格式)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "normal",
                      "warning",
                      "fault",
                      "offline"
                    ],
                    "description": "目标状态",
                    "example": "warning"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "状态更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "设备不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "单独更新设备状态",
        "tags": [
          "设备管理"
        ]
      }
    },
    "/api/equipment/{id}/restore": {
      "post": {
        "description": "将被软删除的设备恢复为可用状态。",
        "operationId": "EquipmentController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "设备ID (UUID格式)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "恢复成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "400": {
            "description": "设备未被删除，无需恢复"
          },
          "401": {
            "description": "未授权，需要登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "设备不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "恢复已删除的设备",
        "tags": [
          "设备管理"
        ]
      }
    },
...
    "/api/statistics/equipment": {
      "get": {
        "description": "获取所有设备的状态统计概览，包括在线、离线和异常设备数量",
        "operationId": "QueryController_getEquipmentOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentOverviewDto"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "设备状态概览",
        "tags": [
          "统计与导出"
        ]
      }
    },
...
  },
  "info": {
    "title": "TDuCargoShipsManagement API",
    "description": "船舶货物管理系统 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEquipmentDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "设备编号（唯一标识）",
            "example": "ENG-001",
            "pattern": "^[A-Z0-9-]+$"
          },
          "deviceName": {
            "type": "string",
            "description": "设备名称",
            "example": "主发动机",
            "minLength": 1,
            "maxLength": 100
          },
          "deviceType": {
            "type": "string",
            "description": "设备类型",
            "example": "主机",
            "minLength": 1,
            "maxLength": 50
          },
          "model": {
            "type": "string",
            "description": "设备型号",
            "example": "MAN B&W 6S50MC",
            "maxLength": 100
          },
          "manufacturer": {
            "type": "string",
            "description": "制造商",
            "example": "MAN Diesel & Turbo",
            "maxLength": 100
          },
          "location": {
            "type": "string",
            "description": "安装位置",
            "example": "机舱中部",
            "maxLength": 100
          },
          "commissionDate": {
            "type": "string",
            "description": "投产日期",
            "example": "2023-01-15",
            "format": "date"
          },
          "description": {
            "type": "string",
            "description": "设备描述",
            "example": "船舶主推进发动机，额定功率9000kW",
            "maxLength": 500
          },
          "status": {
            "type": "string",
            "description": "设备状态",
            "enum": [
              "normal",
              "warning",
              "fault",
              "offline"
            ],
            "example": "normal",
            "default": "normal"
          }
        },
        "required": [
          "deviceId",
          "deviceName",
          "deviceType"
        ]
      },
      "Equipment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "设备唯一ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "deviceId": {
            "type": "string",
            "description": "设备编号（唯一标识）",
            "example": "ENG-001"
          },
          "deviceName": {
            "type": "string",
            "description": "设备名称",
            "example": "左推进电机"
          },
          "deviceType": {
            "type": "string",
            "description": "设备类型",
            "example": "推进电机"
          },
          "model": {
            "type": "string",
            "description": "设备型号",
            "example": "PM-1500-400V"
          },
          "manufacturer": {
            "type": "string",
            "description": "制造商",
            "example": "某电机制造商"
          },
          "location": {
            "type": "string",
            "description": "安装位置",
            "example": "机舱左侧"
          },
          "status": {
            "type": "string",
            "description": "设备状态",
            "enum": [
              "normal",
              "warning",
              "fault",
              "offline"
            ],
            "example": "normal"
          },
          "commissionDate": {
            "format": "date-time",
            "type": "string",
            "description": "投产日期",
            "example": "2024-02-01"
          },
          "description": {
            "type": "string",
            "description": "设备描述",
            "example": "永磁同步电机，额定功率1500kW"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "记录创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "记录更新时间"
          }
        },
        "required": [
          "id",
          "deviceId",
          "deviceName",
          "deviceType",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateEquipmentDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "设备编号（唯一标识）",
            "example": "ENG-001",
            "pattern": "^[A-Z0-9-]+$"
          },
          "deviceName": {
            "type": "string",
            "description": "设备名称",
            "example": "主发动机",
            "minLength": 1,
            "maxLength": 100
          },
          "deviceType": {
            "type": "string",
            "description": "设备类型",
            "example": "主机",
            "minLength": 1,
            "maxLength": 50
          },
          "model": {
            "type": "string",
            "description": "设备型号",
            "example": "MAN B&W 6S50MC",
            "maxLength": 100
          },
          "manufacturer": {
            "type": "string",
            "description": "制造商",
            "example": "MAN Diesel & Turbo",
            "maxLength": 100
          },
          "location": {
            "type": "string",
            "description": "安装位置",
            "example": "机舱中部",
            "maxLength": 100
          },
          "commissionDate": {
            "type": "string",
            "description": "投产日期",
            "example": "2023-01-15",
            "format": "date"
          },
          "description": {
            "type": "string",
            "description": "设备描述",
            "example": "船舶主推进发动机，额定功率9000kW",
            "maxLength": 500
          },
          "status": {
            "type": "string",
            "description": "设备状态",
            "enum": [
              "normal",
              "warning",
              "fault",
              "offline"
            ],
            "example": "normal",
            "default": "normal"
          }
        }
      },
...
      "EquipmentOverviewDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "number",
            "description": "设备总数",
            "example": 50
          },
          "onlineCount": {
            "type": "number",
            "description": "在线设备数（状态为normal或warning）",
            "example": 45
          },
          "offlineCount": {
            "type": "number",
            "description": "离线设备数（状态为offline）",
            "example": 3
          },
          "abnormalCount": {
            "type": "number",
            "description": "异常设备数（有未处理告警）",
            "example": 2
          }
        },
        "required": [
          "totalCount",
          "onlineCount",
          "offlineCount",
          "abnormalCount"
        ]
      },
...
    }
  }
}